requirements.txt
PyPDF2>=2.0.0
pyttsx3>=2.90
SpeechRecognition>=3.8.1
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
streamlit>=1.0.0